[TwigBridge] Add FormLayoutTestCase class
