+++ src/Symfony/Bridge/Twig/Tests/Extension/FormExtensionDivLayoutTest.php
@@ -24,34 +24,6 @@ use Twig\Loader\FilesystemLoader;
 
 class FormExtensionDivLayoutTest extends AbstractDivLayoutTestCase
 {
-    use RuntimeLoaderProvider;
-
-    private FormRenderer $renderer;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $loader = new FilesystemLoader([
-            __DIR__.'/../../Resources/views/Form',
-            __DIR__.'/Fixtures/templates/form',
-        ]);
-
-        $environment = new Environment($loader, ['strict_variables' => true]);
-        $environment->addExtension(new TranslationExtension(new StubTranslator()));
-        $environment->addGlobal('global', '');
-        // the value can be any template that exists
-        $environment->addGlobal('dynamic_template_name', 'child_label');
-        $environment->addExtension(new FormExtension());
-
-        $rendererEngine = new TwigRendererEngine([
-            'form_div_layout.html.twig',
-            'custom_widgets.html.twig',
-        ], $environment);
-        $this->renderer = new FormRenderer($rendererEngine, $this->createMock(CsrfTokenManagerInterface::class));
-        $this->registerTwigRuntimeLoader($environment, $this->renderer);
-    }
-
     public function testThemeBlockInheritanceUsingUse()
     {
         $view = $this->factory
@@ -323,60 +295,6 @@ class FormExtensionDivLayoutTest extends AbstractDivLayoutTestCase
         $this->assertMatchesXpath($html, '/label[@for="name"][@class="my&class required"]/b[.="Bolded label"]');
     }
 
-    protected function renderForm(FormView $view, array $vars = [])
-    {
-        return $this->renderer->renderBlock($view, 'form', $vars);
-    }
-
-    protected function renderLabel(FormView $view, $label = null, array $vars = [])
-    {
-        if (null !== $label) {
-            $vars += ['label' => $label];
-        }
-
-        return $this->renderer->searchAndRenderBlock($view, 'label', $vars);
-    }
-
-    protected function renderHelp(FormView $view)
-    {
-        return $this->renderer->searchAndRenderBlock($view, 'help');
-    }
-
-    protected function renderErrors(FormView $view)
-    {
-        return $this->renderer->searchAndRenderBlock($view, 'errors');
-    }
-
-    protected function renderWidget(FormView $view, array $vars = [])
-    {
-        return $this->renderer->searchAndRenderBlock($view, 'widget', $vars);
-    }
-
-    protected function renderRow(FormView $view, array $vars = [])
-    {
-        return $this->renderer->searchAndRenderBlock($view, 'row', $vars);
-    }
-
-    protected function renderRest(FormView $view, array $vars = [])
-    {
-        return $this->renderer->searchAndRenderBlock($view, 'rest', $vars);
-    }
-
-    protected function renderStart(FormView $view, array $vars = [])
-    {
-        return $this->renderer->renderBlock($view, 'form_start', $vars);
-    }
-
-    protected function renderEnd(FormView $view, array $vars = [])
-    {
-        return $this->renderer->renderBlock($view, 'form_end', $vars);
-    }
-
-    protected function setTheme(FormView $view, array $themes, $useDefaultThemes = true)
-    {
-        $this->renderer->setTheme($view, $themes, $useDefaultThemes);
-    }
-
     public static function themeBlockInheritanceProvider()
     {
         return [
@@ -388,6 +306,39 @@ class FormExtensionDivLayoutTest extends AbstractDivLayoutTestCase
     {
         return [
             [['parent_label.html.twig'], ['child_label.html.twig']],
+        ];
+    }
+
+    protected function getTemplatePaths(): array
+    {
+        return [
+            __DIR__.'/../../Resources/views/Form',
+            __DIR__.'/Fixtures/templates/form',
+        ];
+    }
+
+    protected function getTwigExtensions(): array
+    {
+        return [
+            new TranslationExtension(new StubTranslator()),
+            new FormExtension(),
+        ];
+    }
+
+    protected function getTwigGlobals(): array
+    {
+        return [
+            'global' => '',
+            // the value can be any template that exists
+            'dynamic_template_name' => 'child_label',
+        ];
+    }
+
+    protected function getThemes(): array
+    {
+        return [
+            'form_div_layout.html.twig',
+            'custom_widgets.html.twig',
         ];
     }
 }
